% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{area.s2region}
\alias{area.s2region}
\title{Area of s2region}
\usage{
\method{area}{s2region}(w)
}
\arguments{
\item{w}{Object of class \code{"s2region"}.}
}
\description{
Area of s2region
}
\examples{
region <- s2(radius = 2)
area(region)
region <- s2cap(axis = c(lon=0,lat=0), height = 1, radius = 2)
area(region)
}
