% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{as.data.frame.s2pp}
\alias{as.data.frame.s2pp}
\title{Export coordinates of a spherical point pattern into a data.frame}
\usage{
\method{as.data.frame}{s2pp}(x, ..., format = "long,lat")
}
\arguments{
\item{x}{Spherical point patten of class \code{"s2pp"}.}

\item{...}{Ignored.}

\item{format}{String. Type of coordinates used for return coordinates.
Defaults to \code{"long,lat"}.}
}
\value{
\code{data.frame} with coordinates.
}
\description{
Export coordinates of a spherical point pattern into a data.frame
}
\details{
Return format would usually be a latitude,longitude format (in any
order) by specifying the desired names of the two output columns. Valid
choices for latitude are \code{lat} and \code{latitude}. Valid choices for longitude
are \code{lon}, \code{long}, \code{lng} and \code{longitude}. Alternatively, any order of x,y,z
coordinates are obtained by supplying a commaseparated format like \code{"y,z,x"}.
}
