% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{as.s2cellid}
\alias{as.s2cellid}
\title{Coerce multiple inputs to a s2cellid vector}
\usage{
as.s2cellid(...)
}
\arguments{
\item{...}{Inputs. Currently only objects of class \code{"s2cellid"} are understood.}
}
\value{
Object of class \code{"s2cellid"}.
}
\description{
Coerce multiple inputs to a s2cellid vector
}
