% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{as.s2im.function}
\alias{as.s2im.function}
\title{Convert Function on the Sphere to an Image/Raster on the Sphere}
\usage{
as.s2im.function(f, region = NULL, ..., level = NULL)
}
\arguments{
\item{f}{function of degrees longitude and latitude in that order.}

\item{region}{Object of class \code{s2region} where the function should be evaluated.}

\item{...}{Ignored.}

\item{level}{Integer level of the pixels between 0 (each pixel is huge --
millions of square kilometres on Earth) and 30 (each pixel is tiny --
approximately a square cm on Earth).}
}
\value{
Object of class \code{s2im}.
}
\description{
Convert Function on the Sphere to an Image/Raster on the Sphere
}
\examples{
poly <- s2polygon(cbind(lon = c(-20,-10,10,20,10,-10), lat = c(0, -10, -10, 0, 10, 10)))
f <- function(lon, lat){ sin(lon) }
im <- as.s2im.function(f, poly, level = 6)
plot(im)
}
