% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{as.s2pp.s2grid}
\alias{as.s2pp.s2grid}
\title{Convert a Grid on the Sphere to a Point Pattern on the Sphere}
\usage{
\method{as.s2pp}{s2grid}(X, ...)
}
\arguments{
\item{X}{Object of class \code{s2grid}.}

\item{...}{Ignored.}
}
\value{
Point pattern of class \code{s2pp}. The points are the grid cell
centers.
}
\description{
Convert a Grid on the Sphere to a Point Pattern on the Sphere
}
\examples{
poly <- s2polygon(cbind(lon = c(-20,-10,10,20,10,-10), lat = c(0, -10, -10, 0, 10, 10)))
grid <- s2grid.s2polygon(poly, 5)
pp <- as.s2pp(grid)
}
