% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{as.s2pp.s2im}
\alias{as.s2pp.s2im}
\title{Convert an Image/Raster on the Sphere to a Marked Point Pattern on the Sphere}
\usage{
\method{as.s2pp}{s2im}(X, ...)
}
\arguments{
\item{X}{Object of class \code{s2im}.}

\item{...}{Ignored.}
}
\value{
Marked point pattern of class \code{s2pp}. The points are the image cell
centers and the marks are the corresponding values.
}
\description{
Convert an Image/Raster on the Sphere to a Marked Point Pattern on the Sphere
}
\examples{
poly <- s2polygon(cbind(lon = c(-20,-10,10,20,10,-10), lat = c(0, -10, -10, 0, 10, 10)))
f <- function(lon, lat){ sin(lon) }
im <- as.s2im.function(f, poly, level = 6)
pp <- as.s2pp(im)
plot(pp)
}
