% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dist.R
\name{closepairs.s2pp}
\alias{closepairs.s2pp}
\title{Close Pairs of Points on Sphere}
\usage{
\method{closepairs}{s2pp}(X, rmax, ..., chordal = FALSE)
}
\arguments{
\item{X}{Point pattern on the sphere of class \code{"s2pp"}.}

\item{rmax}{Numeric}

\item{...}{Further arguments passed to \verb{[spatstat.geom::closepairs.pp3()]}.
Notably the argument \code{what} can be convenient.}

\item{chordal}{logical. Cordal distance cutting through the earth if \code{TRUE}.
Otherwise, distance along the surface of the sphere if \code{FALSE}.}
}
\value{
List with info about close pairs of points
}
\description{
Close Pairs of Points on Sphere
}
\seealso{
spatstat.geom::closepairs.
}
