% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{coords.s2pp}
\alias{coords.s2pp}
\title{Extract coordinates of a point pattern on a sphere}
\usage{
\method{coords}{s2pp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"s2pp"}}

\item{...}{Ignored}
}
\value{
a \code{data.frame} with one row for each point, containing the
coordinates.
}
\description{
Extract coordinates of a point pattern on a sphere
}
