% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plotting.R
\name{iplot}
\alias{iplot}
\title{Interactive plotting function}
\usage{
iplot(x, ...)
}
\arguments{
\item{x}{spherical data object to be plotted interactively.}

\item{...}{Additional parameters passed to methods.}
}
\description{
Generic to plot spherical data on an interactive sphere.
}
