% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plotting.R
\name{iplot.s2pp}
\alias{iplot.s2pp}
\title{Plot a spherical point pattern on an interactive sphere with rotation and zoom}
\usage{
\method{iplot}{s2pp}(x, ...)
}
\arguments{
\item{x}{Spherical point pattern of class \code{"s2pp"}.}

\item{...}{Additional parameters passed to \link{iplot.s2polygon} and \code{\link[threejs:globejs]{threejs::globejs()}} to control
the appearance of the plot. For instance \code{img} can be a file path to an image
to wrap on the sphere, and \code{color}, and \code{pointsize} control the appearance of
the plotted points.}
}
\description{
Plot a spherical point pattern on an interactive sphere using \code{\link[threejs:globejs]{threejs::globejs()}}.
}
