% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dist.R
\name{nndist.s2pp}
\alias{nndist.s2pp}
\title{Nearest Neighbour Distance for Points on Sphere}
\usage{
\method{nndist}{s2pp}(X, ..., k = 1, chordal = FALSE)
}
\arguments{
\item{X}{Point pattern on the sphere of class \code{"s2pp"}.}

\item{...}{Further arguments passed to \verb{[spatstat.geom::nndist.pp3()]}.
Only the argument \code{by} can be used by this function.}

\item{k}{Positive integer to calculate the kth nearest neighbour.}

\item{chordal}{logical. Cordal distance cutting through the earth if \code{TRUE}.
Otherwise, distance along the surface of the sphere if \code{FALSE}.}
}
\value{
Numeric vector (if \code{length(k)}=1) or matrix (if \code{length(k)}>1) of distances.
If the additional argument \code{by} is used this may be a \code{data.frame}.
}
\description{
Nearest Neighbour Distance for Points on Sphere
}
\seealso{
spatstat.geom::closepairs.
}
