% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dist.R
\name{pairdist.s2pp}
\alias{pairdist.s2pp}
\title{Pairwise Distances Between Points on Sphere}
\usage{
\method{pairdist}{s2pp}(X, ...)
}
\arguments{
\item{X}{Point pattern on the sphere of class \code{"s2pp"}.}

\item{...}{Ignored.}
}
\value{
Numeric matrix of distances.
}
\description{
Pairwise Distances Between Points on Sphere
}
\seealso{
spatstat.geom::pairdist.
}
