% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{plot.s2im}
\alias{plot.s2im}
\title{Plot Image/Raster on the Sphere}
\usage{
\method{plot}{s2im}(x, ...)
}
\arguments{
\item{x}{Object of class \code{s2im}.}

\item{...}{passed to <plot.s2pp>.}
}
\description{
Plot Image/Raster on the Sphere
}
\examples{
poly <- s2polygon(cbind(lon = c(-20,-10,10,20,10,-10), lat = c(0, -10, -10, 0, 10, 10)))
f <- function(lon, lat){ sin(lon) }
im <- as.s2im.function(f, poly, level = 6)
plot(im)
}
