% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{print.s2cellid}
\alias{print.s2cellid}
\title{Print object of class s2cellid}
\usage{
\method{print}{s2cellid}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"s2cellid"}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly)
}
\description{
Print object of class s2cellid
}
