% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{print.s2im}
\alias{print.s2im}
\title{Print Method for Image/Raster on the Sphere}
\usage{
\method{print}{s2im}(x, ...)
}
\arguments{
\item{x}{Object of class \code{s2pp}}

\item{...}{Ignored.}
}
\description{
Print Method for Image/Raster on the Sphere
}
