% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2index.R
\name{print.s2index}
\alias{print.s2index}
\title{Print object of class s2index}
\usage{
\method{print}{s2index}(x, ...)
}
\arguments{
\item{x}{Object of class \code{s2index}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly).
}
\description{
Print object of class s2index.
}
