% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2point.R
\name{print.s2point}
\alias{print.s2point}
\title{Print object of class s2point}
\usage{
\method{print}{s2point}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"s2point"}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly)
}
\description{
Print object of class s2point
}
