% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{print.s2pp}
\alias{print.s2pp}
\title{Print object of class s2pp}
\usage{
\method{print}{s2pp}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"s2pp"}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly)
}
\description{
Print object of class s2pp
}
