% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{print.s2region}
\alias{print.s2region}
\title{Print object of class s2region}
\usage{
\method{print}{s2region}(x, ...)
}
\arguments{
\item{x}{Object of class \code{"s2region"}.}

\item{...}{Ignored.}
}
\value{
NULL (invisibly)
}
\description{
Print object of class s2region
}
