% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2}
\alias{s2}
\title{Generate a sphere}
\usage{
s2(radius = 1, unitname = NULL, ...)
}
\arguments{
\item{radius}{Numeric. Defaults to 1. For convenience this can also be an
object of class \code{"s2region"} from where the radius and unitname then will be
extracted.}

\item{unitname}{Optional. Name of unit of length. Either a single character
string, or a vector of two character strings giving the singular and plural
forms, respectively.}

\item{...}{Ignored.}
}
\value{
Object of class \code{"s2"} and super class \code{"s2region"}.
}
\description{
Generate a sphere
}
