% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2cap}
\alias{s2cap}
\title{Define spherical cap}
\usage{
s2cap(axis, height, dist, ..., simplify = TRUE)
}
\arguments{
\item{axis}{Axis (or centre) of the cap in a format acceptable to \code{\link{s2point}}.}

\item{height}{Numeric of length one defining the height of the cap. Should be
positive and at most twice the radius of the sphere. Not compatible with
\code{dist}.}

\item{dist}{Numeric of length one defining the extend of the cap. All points
on the sphere within distance \code{dist} of the point defined by
\code{axis} are contained in the cap. Not compatible with \code{height}.}

\item{...}{
  Arguments passed on to \code{\link[=s2]{s2}}
  \describe{
    \item{\code{radius}}{Numeric. Defaults to 1. For convenience this can also be an
object of class \code{"s2region"} from where the radius and unitname then will be
extracted.}
    \item{\code{unitname}}{Optional. Name of unit of length. Either a single character
string, or a vector of two character strings giving the singular and plural
forms, respectively.}
  }}

\item{simplify}{Logical. Whether to simplify to full sphere if the cap
extends over the full sphere.}
}
\value{
Object of class \code{"s2cap"} and \code{"s2region"}.
}
\description{
Define spherical cap
}
