% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{s2cell}
\alias{s2cell}
\title{Make a list of s2cells}
\usage{
s2cell(x)
}
\arguments{
\item{x}{Input to create cells from. Currently only a vector of s2cellids
are supported.}
}
\description{
Make a list of s2cells
}
