% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{s2cellid}
\alias{s2cellid}
\title{Make s2cellids from points on the sphere}
\usage{
s2cellid(x, level = 30)
}
\arguments{
\item{x}{Points on the sphere in a standard format.}

\item{level}{Integer between 0 and 30 (incl).}
}
\value{
A vector of doubles representing s2cellids with the additional class \code{s2cellid}.
}
\description{
Create a vector of s2cellids corresponding to the cells at the given level
containing the given points. The default level (30) corresponds to leaf
cells (finest/maximal level).
}
