% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{s2cellstring}
\alias{s2cellstring}
\title{Retreive string representation of a s2cell}
\usage{
s2cellstring(x, level = 30L)
}
\arguments{
\item{x}{Input to retrieve string representations from.
Current options are a vector of s2cellid tokens, a list of s2cells, or a
three-column matrix representing points on the sphere.}

\item{level}{Integer between 0 and 30 (incl.) to specify the desired s2cell
level. Defaults to 30 for point data and is ignored for s2cell(id) data.}
}
\description{
Retreive string representation of a s2cell
}
