% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2contains}
\alias{s2contains}
\title{Test whether points are inside s2region}
\usage{
s2contains(coords, region, options = s2::s2_options(model = "closed"))
}
\arguments{
\item{coords}{\code{data.frame}, \code{matrix} or \code{list}. Object containing spherical
coordinates of the points. This must be either longitude and latitude in
degrees or 3D Cartesian coordinates. In the former case two named columns
or list entries are required. In the latter three colums or list entries
are required and assumed to be in x, y, z order (disregarding any names).}

\item{region}{Region on the sphere where the points occur. Object of class
\code{"s2region"}.}

\item{options}{List of class \code{"s2_options"} to control how containment is
defined for polygonal regions. See \code{\link[s2:s2_options]{s2::s2_options}} for available options
and details.}
}
\value{
Logical vector of same length as the number of rows in \code{coords}
}
\description{
Test whether points are inside s2region
}
