% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{s2coords}
\alias{s2coords}
\title{Extract Coordinates of a Point Pattern on a Sphere}
\usage{
s2coords(x)
}
\arguments{
\item{x}{Object of class \code{"s2pp"}}
}
\value{
a \code{data.frame} with one row for each point, containing the
coordinates.
}
\description{
Extract Coordinates of a Point Pattern on a Sphere
}
