% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{s2covering}
\alias{s2covering}
\title{Approximate a region on the sphere by a covering of s2cells}
\usage{
s2covering(x, max_cells = 8, min_level = 0, max_level = 30, interior = FALSE)
}
\arguments{
\item{x}{Region to cover. Currently it must be a polygon, cap or full sphere.}

\item{max_cells}{Positive integer. Maximal number of cells to use in the
covering.}

\item{min_level}{Integer between 0 and 30 specifying the lowest cell level to
use. Must be less than or equal to \code{max_level}.}

\item{max_level}{Integer between 0 and 30 specifying the highest cell level to
use. Must be greater than or equal to \code{min_level}.}

\item{interior}{Logical to get an interior covering.}
}
\value{
A vector of s2cellids (of class \code{s2cellid}).
}
\description{
Approximate a region on the sphere by a (possibly interior) covering of
s2cells.
}
\examples{
# Covering of entire sphere at level 1
s2covering(s2earth(), min_level = 1, max_level = 1)

}
