% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2dist.R
\name{s2dist}
\alias{s2dist}
\title{Distance on sphere}
\usage{
s2dist(x, y)
}
\arguments{
\item{x}{One or more points in a format acceptable to \code{\link[globe:ensure3d]{globe::ensurelonlat()}}.}

\item{y}{One or more points in a format acceptable to \code{\link[globe:ensure3d]{globe::ensurelonlat()}}.}
}
\value{
Numeric vector of distances.
}
\description{
Distance on sphere
}
\details{
If both \code{x} and \code{y} contain more than one point they must contain
the same number of points and the distances between corresponding points are
returned.
}
