% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2earth}
\alias{s2earth}
\title{Sphere approximating earth}
\usage{
s2earth()
}
\value{
sphere of class \code{"s2"}.
}
\description{
Sphere approximating earth
}
\examples{
earth <- s2earth()
area(earth)
}
