% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2Gest.R
\name{s2Gest}
\alias{s2Gest}
\title{G-function for spherical point pattern}
\usage{
s2Gest(X, r = NULL, rmax = NULL, breaks = NULL, correction = NULL)
}
\arguments{
\item{X}{Spherical point pattern of class \code{s2pp}}

\item{r}{Optional. Vector of values for the argument \eqn{r} at which
\eqn{G(r)} should be evaluated. Users are advised \emph{not} to specify
this argument; there is a sensible default. If necessary, specify
\code{rmax}.}

\item{rmax}{Optional. Maximum desired value of the argument \eqn{r}.}

\item{breaks}{This argument is for internal use only.}

\item{correction}{Optional. A character vector containing any selection of
the options \code{"none"} and \code{"border"} (and possibly others in the
future).}
}
\value{
An object of class \code{"fv"}, see \code{\link{fv.object}}, which
can be plotted directly using \code{\link{plot.fv}}.

Essentially a data frame containing columns \item{r}{the vector of values
of the argument \eqn{r} at which the function \eqn{G} has been estimated }
\item{theo}{the theoretical value \eqn{G(r) = 2 \pi (1-cos(r))}{G(r) = 2 *
  pi * (1 - cos(r))} for a stationary Poisson process (where r is in
radians)} together with columns named \code{"none"}, and/or
\code{"border"}, according to the selected edge corrections. These columns
contain estimates of the function \eqn{G(r)} obtained by the edge
corrections named.
}
\description{
G-function for spherical point pattern
}
\examples{
## Random poins on earth
X <- s2runif(200, region = s2earth())
## Estimate of G function up to distance 2000 km
G_X <- s2Gest(X, rmax = 2000)

}
