% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{s2grid.s2polygon}
\alias{s2grid.s2polygon}
\title{Make a grid over spherical polygonal region}
\usage{
s2grid.s2polygon(x, level, ...)
}
\arguments{
\item{x}{Polygon on the sphere of class \code{s2polygon}.}

\item{level}{Integer level of the pixels between 0 (each pixel is huge --
millions of square kilometres on Earth) and 30 (each pixel is tiny --
approximately a square cm on Earth).}

\item{...}{Ignored.}
}
\value{
Object of class "s2grid", which is just a vector of unique ids of
the cells making up the grid.
}
\description{
Make a grid over spherical polygonal region
}
\examples{
poly <- s2polygon(cbind(lon = c(-20,-10,10,20,10,-10), lat = c(0, -10, -10, 0, 10, 10)))
grid <- s2grid.s2polygon(poly, 5)
length(grid)
}
