% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2im.R
\name{s2im}
\alias{s2im}
\title{Make an Image/Raster Object on the Sphere}
\usage{
s2im(grid, values)
}
\arguments{
\item{grid}{An \code{s2grid} defining the cells/pixels of the image.}

\item{values}{A \code{numeric} of same length as \code{grid} containing the image
value at each cell.}
}
\value{
Object of class \code{s2im}.
}
\description{
Make an Image/Raster Object on the Sphere
}
\examples{
poly <- s2polygon(cbind(lon = c(-20,-10,10,20,10,-10), lat = c(0, -10, -10, 0, 10, 10)))
grid <- s2grid.s2polygon(poly, 5)
im <- s2im(grid, 1:length(grid))
plot(im)
}
