% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2intersect}
\alias{s2intersect}
\alias{s2union}
\title{Intersection or union of two spherical regions}
\usage{
s2intersect(x, y)
}
\arguments{
\item{x, y}{Region on the sphere of class \code{"s2region"}.}
}
\value{
Region on the sphere of class \code{"s2region"}.
}
\description{
Intersection or union of two objects of class \code{"s2region"}.
}
\examples{
loop1 <- cbind(lon = c(0,60,60,0), lat = c(-40,-40,40,40))
loop2 <- cbind(lon = c(30,90,90,30), lat = c(-40,-40,40,40))
poly1 <- s2polygon(loop1)
poly2 <- s2polygon(list(loop2))
poly3 <- s2intersect(poly1, poly2)
poly4 <- s2union(poly1, poly2)
plot(poly1, eye = c(lon=45, lat=0), col = "red", eps = pi/100)
plot(poly2, col = "green", add = TRUE, eps = pi/100)
plot(poly3, col = "blue", add = TRUE, eps = pi/100)
plot(poly4, col = "cyan", eps = pi/100)
}
