% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2cell.R
\name{s2level}
\alias{s2level}
\title{Extract s2 cell level from data}
\usage{
s2level(x, simplify = FALSE)
}
\arguments{
\item{x}{object inheriting from \code{s2cellid}}

\item{simplify}{logical to return a single value if all levels are equal.}
}
\value{
Integer vector.
}
\description{
Extract s2 cell level from data
}
\examples{
x <- rbind(c(1,0,0), c(0,1,0))
id <- s2cellid(x)
s2level(id)
s2level(id, simplify = TRUE)

id <- s2cellid(x, c(1,30))
s2level(id)
}
