% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2index.R
\name{s2lookup}
\alias{s2lookup}
\title{Look up IDs in a given s2index (beta version)}
\usage{
s2lookup(id, index)
}
\arguments{
\item{id}{Cell ids in the s2cellid format.}

\item{index}{Index where the ids should be looked up.}
}
\value{
Vector of integers matching the s2cellids.
}
\description{
Look up given s2cellids in a s2index and return the matching indices (without
replication). Expect changes to the API or even complete removal
of the function. Only for testing at the moment!
}
