% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2looplist}
\alias{s2looplist}
\title{Create a list of loops on the sphere}
\usage{
s2looplist(...)
}
\arguments{
\item{...}{Input to create the loops from. Each input should be a matrix or
data.frame with the coordinates of each vertex of the loop in each row. A
\code{s2polygon} or \code{s2looplist}.}
}
\value{
A list of loops with class \code{s2looplist}.
}
\description{
Create a list of loops on the sphere.
}
\examples{
loop1 <- cbind(lon = c(0,60,60,0), lat = c(-40,-40,40,40))
loop2 <- cbind(lon = c(30,90,90,30), lat = c(-40,-40,40,40))
x <- s2looplist(loop1, loop2)
y <- s2looplist(x, loop1)
z <- s2looplist(x, s2polygon(loop1))
}
