% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2place.R
\name{s2place}
\alias{s2place}
\title{Convert place name to location on sphere}
\usage{
s2place(placename, online = TRUE, verbose = FALSE)
}
\arguments{
\item{placename}{What to find.}

\item{online}{Logical. Should an online lookup be performed if needed?}

\item{verbose}{Logical. Should verbose output be produced?}
}
\description{
Convert place name to location on sphere
}
