% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2plotting.R
\name{s2png}
\alias{s2png}
\title{Start PNG plotting device for spherical objects}
\usage{
s2png(filename, ..., init = FALSE)
}
\arguments{
\item{filename}{Name of output file. A temporary file will be generated if it
none is given.}

\item{...}{
  Arguments passed on to \code{\link[grDevices:png]{grDevices::png}}
  \describe{
    \item{\code{width}}{the width of the device.}
    \item{\code{height}}{the height of the device.}
    \item{\code{units}}{The units in which \code{height} and \code{width} are
    given.  Can be \code{px} (pixels, the default), \code{in} (inches),
    \code{cm} or \code{mm}.}
    \item{\code{pointsize}}{the default pointsize of plotted text, interpreted as
    big points (1/72 inch) at \code{res} ppi.}
    \item{\code{bg}}{the initial background colour: can be overridden by setting
    par("bg").}
    \item{\code{res}}{The nominal resolution in ppi which will be recorded in the
    bitmap file, if a positive integer.  Also used for \code{units}
    other than the default, and to convert points to pixels.}
    \item{\code{type}}{character string, one of \code{"Xlib"} or \code{"quartz"}
    (some macOS builds) or \code{"cairo"}.  The latter will only be
    available if the system was compiled with support for cairo --
    otherwise \code{"Xlib"} will be used.  The default is set by
    \code{\link{getOption}("bitmapType")} -- the \sQuote{out of the box}
    default is \code{"quartz"} or \code{"cairo"} where available,
    otherwise \code{"Xlib"}.}
    \item{\code{antialias}}{for \code{type = "cairo"}, giving the type of
    anti-aliasing (if any) to be used for fonts and lines (but not
    fills). See \code{\link[grDevices]{X11}}.  The default is set by
    \code{\link[grDevices]{X11.options}}.  Also for \code{type = "quartz"}, where
    antialiasing is used unless \code{antialias = "none"}.}
  }}

\item{init}{Logical to initialize plot with a blank plot with
\code{xlim = c(-180,180)} and \code{ylim = c(-90,90)}.}
}
\value{
Name of output file.
}
\description{
This command starts a PNG plotting device and sets appropriate parameters for
plotting spherical objects. The device must be closed by the user by calling
\code{grDevices::dev.off()} or similar when no more graphics will be added.
}
\examples{
img <- s2png()
globe::flatearth()
dev.off()
file.remove(img)

}
