% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2point.R
\name{s2point}
\alias{s2point}
\title{Single point on the (unit) sphere}
\usage{
s2point(...)
}
\arguments{
\item{...}{Input to interpret as a point on the sphere. Either in terms of
three Eucledian coordinates (x,y,z) or two spherical coordinates
(latitude,longitude). See Details.}
}
\value{
Object of class \code{"s2pp"}.
}
\description{
Make an object of class \code{"s2point"} representing a single point
on the unit sphere.
}
\details{
For three Eucledian coordinates the following type of inputs are
recognised:
\itemize{
\item Three inputs with a single real number (optionally named \code{x}, \code{y} and \code{z}).
\item A list or data.frame with three real numbers as elements (optionally named \code{x}, \code{y} and \code{z}).
\item A single vector or matrix with three real numbers.
}

For two spherical coordinates a named input is required. The names can either
be given as argument names of the function or as the names of the provided
input. Input latitude is partially matched to \code{"latitude"} and input longitude
is either partially matched to \code{longitude} or exactly matched to \code{lng}. In all
cases measurements are in degrees and should be in \eqn{[-90,90]} (latitude)
and \eqn{[-180,180]} (longitude).
The following type of inputs are recognised:
\itemize{
\item Two named inputs with a single real number.
\item A list or data.frame with two named elements containing a real number each.
\item A single vector with two named elements containing a real number each.
\item A two column matrix with a single row of real numbers and a \code{colnames}
attribute.
}
}
\examples{
longitude <- c(90)
latitude <- c(45)
xyz <- s2point(lon = longitude, lat = latitude)
x <- xyz[1]
y <- xyz[2]
z <- xyz[3]
xyz2 <- s2point(x=x, y=y, z=z)

}
