% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2polygon}
\alias{s2polygon}
\title{Create a s2polygon}
\usage{
s2polygon(x, ..., oriented = FALSE, check = TRUE, options = NULL)
}
\arguments{
\item{x}{List of loops (also called rings) to join as a polygon.}

\item{...}{
  Arguments passed on to \code{\link[=s2]{s2}}
  \describe{
    \item{\code{radius}}{Numeric. Defaults to 1. For convenience this can also be an
object of class \code{"s2region"} from where the radius and unitname then will be
extracted.}
    \item{\code{unitname}}{Optional. Name of unit of length. Either a single character
string, or a vector of two character strings giving the singular and plural
forms, respectively.}
  }}

\item{oriented}{logical, which defaults to FALSE. Set to TRUE if polygon ring
directions are known to be correct such that exterior rings are counter
clockwise and interior rings are clockwise.}

\item{check}{logical, which defaults to TRUE. Throws an error if the polygon is invalid.}

\item{options}{List of class \code{"s2_options"} to control how the polygon is
assembled. See \code{\link[s2:s2_options]{s2::s2_options()}} for available options and details.}
}
\description{
Create a s2polygon (object of class \code{"s2polygon"}).
}
