% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{s2pp}
\alias{s2pp}
\title{Spherical point pattern}
\usage{
s2pp(coords, region = NULL, marks = NULL, ..., check = TRUE)
}
\arguments{
\item{coords}{\code{data.frame}, \code{matrix} or \code{list}. Object containing spherical
coordinates of the points. This must be either longitude and latitude in
degrees or 3D Cartesian coordinates. In the former case two named columns
or list entries are required. In the latter three colums or list entries
are required and assumed to be in x, y, z order (disregarding any names).}

\item{region}{Region on the sphere where the points occur. Object of class
\code{"s2region"}.}

\item{marks}{Marks attached to the points.}

\item{...}{
  Arguments passed on to \code{\link[=s2]{s2}}
  \describe{
    \item{\code{radius}}{Numeric. Defaults to 1. For convenience this can also be an
object of class \code{"s2region"} from where the radius and unitname then will be
extracted.}
    \item{\code{unitname}}{Optional. Name of unit of length. Either a single character
string, or a vector of two character strings giving the singular and plural
forms, respectively.}
  }}

\item{check}{Logical. Check that points actually are inside the provided
\code{region}.}
}
\value{
Object of class \code{"s2pp"}.
}
\description{
Make an object of class \code{"s2pp"} representing a point pattern
over a region on the sphere (of class \code{"s2region"}).
}
\examples{
long <- c(90, 180, -90)
lat <- c(45, 0, -45)
coords <- data.frame(long, lat)
X <- s2pp(coords)

}
