% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2radius}
\alias{s2radius}
\title{Radius of sphere}
\usage{
s2radius(x, format = c("numeric", "string", "list"))
}
\arguments{
\item{x}{Object of class \code{"s2region"}.}

\item{format}{String to return the result as a numeric (\code{"numeric"}), a string
with unitname (\code{"string"}) or a list with entries \code{radius} and \code{units}
(\code{"list"})}
}
\value{
numeric (default), string or list with the radius of the
underlying sphere of the \code{"s2region"} and for string/list also the unitname.
}
\description{
Radius of sphere
}
