% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2region}
\alias{s2region}
\title{Extract spherical region}
\usage{
s2region(x)
}
\arguments{
\item{x}{Object of class \code{"s2pp"}.}
}
\value{
object of class \code{"s2region"}.
}
\description{
Extract spherical region
}
