% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{spatstat.sphere}
\alias{spatstat.sphere}
\alias{spatstat.sphere-package}
\title{Package documentation for spatstat.sphere}
\description{
The package \code{spatstat.sphere} extends the \code{spatstat} package to work with
point pattens on the sphere.
}
