% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{[.s2pp}
\alias{[.s2pp}
\title{Extract a subset of a point pattern on a sphere.}
\usage{
\method{[}{s2pp}(x, i, j, drop, ..., clip = FALSE)
}
\arguments{
\item{x}{object of class \code{"s2pp"}.}

\item{i}{Subset index. A valid subset index in the usual R sense, indicating
which points should be retained.}

\item{j}{Ignored. (Required for compatibility with the generic function.)}

\item{drop}{Ignored. (Required for compatibility with the generic function.)}

\item{\dots}{Ignored. (Required for compatibility with the generic function.)}

\item{clip}{Logical value indicating how to form the s2region of the
resulting point pattern, when i is a window. If clip=FALSE (the default), the
result has s2region equal to i. If clip=TRUE, the resulting s2region is the
intersection between the s2region of x and the s2region i.}
}
\value{
object of class \code{"s2pp"}.
}
\description{
Extract a subset of a point pattern on a sphere. Extraction of a subset has
the effect of thinning the points.
}
\examples{

X <- s2pp(data.frame(lat = c(-45,0,45), long = c(-10,0,160)))
X[1:2]

}
