% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2pp.R
\name{superimpose.s2pp}
\alias{superimpose.s2pp}
\title{Superimpose several point patterns on the sphere}
\usage{
\method{superimpose}{s2pp}(..., region = NULL, check = TRUE)
}
\arguments{
\item{...}{Any number of arguments, each of which represents a point pattern
on the sphere.}

\item{region}{Optional. Object of class \code{"s2region"} determining the region
for the resulting pattern.}

\item{check}{Logical value passed to \code{\link[=s2pp]{s2pp()}} determining whether to check
the geometrical validity of the resulting pattern.}
}
\value{
Point pattern on the sphere of class \code{"s2pp"}.
}
\description{
Superimpose several point patterns on the sphere
}
\examples{
p1 <- s2runif(10)
p2 <- s2runif(20)
s1 <- superimpose(p1, p2)
}
