% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2contains}
\alias{s2contains}
\title{Test whether points are inside s2region}
\usage{
s2contains(coords, region, approx = TRUE)
}
\arguments{
\item{coords}{\code{data.frame}, \code{matrix} or \code{list}. Object containing spherical
coordinates of the points. This must be either longitude and latitude in
degrees or 3D Cartesian coordinates. In the former case two named columns
or list entries are required. In the latter three colums or list entries
are required and assumed to be in x, y, z order (disregarding any names).}

\item{region}{Region on the sphere where the points occur. Object of class
\code{"s2region"}.}

\item{approx}{Logical to use approximate containment, which allows points
very slightly outside a polygonal region, but also ensures that points on the
border of a polygon are always accepted. Does not affect spherical caps at
the moment.}
}
\value{
Logical vector of same length as the number of rows in \code{coords}
}
\description{
Test whether points are inside s2region
}
