% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2region.R
\name{s2lonlatbox}
\alias{s2lonlatbox}
\alias{s2lonlatbox.s2region}
\alias{s2lonlatbox.default}
\alias{s2lonlatbox.s2_geography}
\title{Generic function to extract or construct a box in longitude and latitude coordinates}
\usage{
s2lonlatbox(...)

\method{s2lonlatbox}{s2region}(x, ...)

\method{s2lonlatbox}{default}(lon = c(-180, 180), lat = c(-90, 90), ...)

\method{s2lonlatbox}{s2_geography}(x, ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=s2]{s2}}, \code{\link[=s2]{s2}}
  \describe{
    \item{\code{radius}}{Numeric. Defaults to 1. For convenience this can also be an
object of class \code{"s2region"} from where the radius and unitname then will be
extracted.}
    \item{\code{unitname}}{Optional. Name of unit of length. Either a single character
string, or a vector of two character strings giving the singular and plural
forms, respectively.}
  }}

\item{x}{object of class \code{"s2_geography"}.}

\item{lon}{Numeric of length 2 specifying longitude interval in degrees. For
boxes not crossing the 180 degree meridian the input should be such that
\verb{-180 <= lon[1] < lon[2] <= 180}. For boxes crossing the 180 degree meridian
the input shoud be such that \verb{-180 <= lon[2] < lon[1] <= 180}.}

\item{lat}{Numeric of length 2 with \verb{-90 <= lat[1] < lat[2] <= 90} specifying
latitude interval in degrees.}
}
\value{
Box in longitude and latitude coordinates (of class \code{"s2lonlatbox"}).
}
\description{
Generic function to extract or construct a box in longitude and latitude coordinates.
}
\section{Methods (by class)}{
\itemize{
\item \code{s2lonlatbox(s2region)}: Extraction method for \code{"s2region"}.

\item \code{s2lonlatbox(default)}: Default method to generate \code{"s2lonlatbox"} from a
longitude and latitude interval.

\item \code{s2lonlatbox(s2_geography)}: Method to generate \code{"s2lonlatbox"} from an object of
class \code{"s2_geography"}.

}}
\examples{
s2lonlatbox(lon = c(-30, 30), lat = c(45, 90))
## Very large box crossing the 180th meridian
big <- s2lonlatbox(lon = c(1, -1), lat = c(-90, 90))
# area(big) # Not implemented yet
## Very small box **not** crossing the 180th meridian
small <- s2lonlatbox(lon = c(-1, 1), lat = c(-90, 90))
# area(small) # Not implemented yet

}
\seealso{
s2lonlatbox.s2region s2lonlatbox.default
}
